css-in-js loader icon
