+++ src/loader/Loader.unit.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow, mount } from 'enzyme'
 
-import Loader from 'loader'
+import Loader from './Loader'
 import { CircleIcon } from 'icon/circleIcon'
 import { CheckIcon } from 'icon/checkIcon'
