+++ src/loader/story.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { withKnobs, boolean, number } from '@storybook/addon-knobs'
 
-import Loader from 'loader'
+import Loader from './index'
 
 const stories = storiesOf('Loader', module)
 stories.addDecorator(withKnobs)
